requirements.txt
alabaster==0.7.13
Babel==2.11.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
coloredlogs==15.0.1
coverage==7.0.5
cryptography==39.0.0
deprecation==2.1.0
docutils==0.19
humanfriendly==10.0
idna==3.4
imageio==2.22.4
imagesize==1.4.1
img2pdf==0.4.4
importlib-metadata==6.0.0
injector==0.20.1
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
networkx==2.8.8
numpy==1.23.4
ocrmypdf==14.0.2
opencv-python==4.6.0.66
packaging==21.3
parameterized==0.8.1
pdfminer.six==20221105
pikepdf==6.2.8.post1
Pillow==9.3.0
pluggy==1.0.0
pycparser==2.21
Pygments==2.14.0
PyMuPDF==1.21.1
pyparsing==3.0.9
PySide6==6.4.1
PySide6-Addons==6.4.1
PySide6-Essentials==6.4.1
pytesseract==0.3.10
pytz==2022.7.1
PyWavelets==1.4.1
reportlab==3.6.12
requests==2.28.2
scikit-image==0.19.3
scipy==1.9.3
shiboken6==6.4.1
snowballstemmer==2.2.0
sphinx==6.1.3
sphinxcontrib-applehelp==1.0.3
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
tifffile==2022.10.10
tqdm==4.64.1
typing-extensions==4.4.0
urllib3==1.26.14
zipp==3.11.0
